requirements.txt
streamlit
streamlit-drawable-canvas
Pillow
google-api-python-client
google-auth-httplib2
google-auth-oauthlib